+++ package.json
@@ -12,6 +12,8 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "browserify-zlib": "^0.2.0",
+    "ethers": "^6.1.0",
     "framer-motion": "^7.6.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
