+++ src/components/layouts/LandingLayout.jsx
@@ -1,0 +1,18 @@
+import React from "react";
+import { Flex } from "@chakra-ui/react";
+import Header from "../sections/Header";
+
+export default function LandingLayout(props) {
+  return (
+    <Flex
+      direction="column"
+      align="center"
+      maxW={{ xl: "1300px" }}
+      m="0 auto"
+      {...props}
+    >
+      <Header />
+      {props.children}
+    </Flex>
+  );
+}
