+++ src/pages/Landing.js
@@ -1,19 +1,18 @@
 import React, { useEffect } from "react";
-import Logo from '../components/ui/awd.gif'
+import bgLogo from "../components/ui/bgLogo.png";
 import Hero from "../components/sections/Hero";
 import LandingLayout from "../components/layouts/LandingLayout";
 
-
 export default function Landing() {
   useEffect(() => {
-    document.title = 'Perped - A Decentralised Perpetuals Protocol';
+    document.title = "Perped - A Decentralised Perpetuals Protocol";
   }, []);
   return (
     <LandingLayout>
       <Hero
-        title="A Decentralised Perpetuals Protocol"
-        subtitle="COMING SOON"
-        image= { Logo }
+        title="Trade in Perpetuity"
+        subtitle="The First True On-Chain Perpetuals Derivative Platform For Real World Assets"
+        image={bgLogo}
       />
     </LandingLayout>
   );
